feat(compute): Move policy logic to domain/compute
