+++ agent/tests/integration/test_arkhai.py
@@ -29,7 +29,7 @@ def test_torch_import():
 def test_torch_arkhai_seller_import():
     """Test that torch_arkhai_seller module can be imported."""
     try:
-        from app.policies import torch_arkhai_seller
+        import domain.compute.agent.app.policy.torch_arkhai_seller as torch_arkhai_seller
         assert torch_arkhai_seller is not None
     except ImportError as e:
         pytest.fail(f"Failed to import torch_arkhai_seller: {e}")
@@ -51,7 +51,7 @@ def test_observation_builder_basic():
 def test_action_extraction_basic():
     """Test that action extraction works with mock model output."""
     import torch
-    from app.policies.arkhai_common import extract_actions_from_logits as _extract_actions_from_logits
+    from domain.compute.agent.app.policy.arkhai_common import extract_actions_from_logits as _extract_actions_from_logits
 
     # Create mock output (action_logits, values)
     action_logits = torch.randn(1, 11)  # 9 price + 2 sell
@@ -71,11 +71,11 @@ def test_action_extraction_basic():
 
 def test_model_path_environment_variable(monkeypatch):
     """Test that ARKHAI_SELLER_MODEL_PATH environment variable is respected."""
-    from app.policies.arkhai_common import _MODEL_CACHE
+    from domain.compute.agent.app.policy.arkhai_common import _MODEL_CACHE
     # Clear cache to ensure fresh load attempt
     _MODEL_CACHE.clear()
 
-    from app.policies.torch_arkhai_seller import _get_model
+    from domain.compute.agent.app.policy.torch_arkhai_seller import _get_model
 
     # Set environment variable to a test path
     test_path = "/tmp/test_arkhai_model.pt"
