+++ domain/compute/agent/app/policy/store.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import logging
 from typing import Any
 
-from core.agent.app.policy.store import PolicyStore
 from core.agent.app.schema.pydantic_models import (
     Action as DomainAction,
     ActionType as DomainActionType,
