+++ domain/compute/agent/app/policy/torch_arkhai_seller.py
@@ -1,8 +1,9 @@
-"""Arkhai buyer policy adapter backed by upstream puffer checkpoints.
+"""Arkhai seller policy adapter backed by upstream puffer checkpoints.
 
-Mirror of the seller adapter with buyer-specific action thresholds.
-Buyers are more price-sensitive: their reward is profit margin
-(base_price - negotiated_price) * duration, so lower prices are preferred.
+This policy intentionally keeps local logic thin:
+- Observation layout follows upstream Arkhai ordering (via arkhai_common).
+- Model loading expects puffer `state_dict` checkpoints.
+- Action mapping remains local (domain-level accept/counter/reject).
 """
 from __future__ import annotations
 
@@ -10,7 +11,7 @@ import logging
 from pathlib import Path
 from typing import Any, Optional
 
-from app.policies.arkhai_common import (
+from domain.compute.agent.app.policy.arkhai_common import (
     build_action_parameters,
     build_arkhai_observation,
     detect_agent_role,
@@ -30,21 +31,21 @@ from core.agent.app.schema.pydantic_models import (
     MakeOfferEvent,
 )
 from core.agent.app.utils.config import CONFIG
-from app.utils.validation import extract_resources_from_make_offer_event
+from core.agent.app.utils.validation import extract_resources_from_make_offer_event
 
 logger = logging.getLogger(__name__)
 
-_DEFAULT_MODEL_PATH = Path(__file__).resolve().parent / "models" / "arkhai_buyer.pt"
+_DEFAULT_MODEL_PATH = Path(__file__).resolve().parent / "models" / "arkhai_seller.pt"
 
 
 def _get_model(obs_dim_val: int) -> Optional[Any]:
-    """Lazily load the Arkhai buyer model checkpoint."""
-    return get_model("ARKHAI_BUYER_MODEL_PATH", _DEFAULT_MODEL_PATH, obs_dim_val)
+    """Lazily load the Arkhai seller model checkpoint."""
+    return get_model("ARKHAI_SELLER_MODEL_PATH", _DEFAULT_MODEL_PATH, obs_dim_val)
 
 
-@policy_callable("mo.action.torch_arkhai_buyer")
-async def mo_action_torch_arkhai_buyer(context: DecisionContext) -> DomainAction | None:
-    """Model-based Arkhai buyer policy for MakeOffer events."""
+@policy_callable("mo.action.torch_arkhai_seller")
+async def mo_action_torch_arkhai_seller(context: DecisionContext) -> DomainAction | None:
+    """Model-based Arkhai seller policy for MakeOffer events."""
     if not isinstance(context.event, MakeOfferEvent):
         return None
 
@@ -68,24 +69,24 @@ async def mo_action_torch_arkhai_buyer(context: DecisionContext) -> DomainAction
                         },
                     )
             except Exception as exc:
-                logger.warning("[ARKHAI BUYER POLICY] Failed to validate portfolio: %s", exc)
+                logger.warning("[ARKHAI SELLER POLICY] Failed to validate portfolio: %s", exc)
 
     node_types = parse_node_types()
     model = _get_model(obs_dim(node_types))
     if model is None or torch is None:
-        logger.warning("[ARKHAI BUYER POLICY] Model unavailable; returning None")
+        logger.warning("[ARKHAI SELLER POLICY] Model unavailable; returning None")
         return None
 
     observation = build_arkhai_observation(context, offer_resource, demand_resource, order)
     if observation is None:
-        logger.warning("[ARKHAI BUYER POLICY] Failed to build observation; returning None")
+        logger.warning("[ARKHAI SELLER POLICY] Failed to build observation; returning None")
         return None
 
     try:
         with torch.no_grad():
             output = model(observation)
     except Exception as exc:
-        logger.error("[ARKHAI BUYER POLICY] Inference failed: %s", exc)
+        logger.error("[ARKHAI SELLER POLICY] Inference failed: %s", exc)
         return None
 
     price_idx, sell_flag = extract_actions_from_logits(output)
@@ -94,23 +95,21 @@ async def mo_action_torch_arkhai_buyer(context: DecisionContext) -> DomainAction
         offer_resource, demand_resource, order, CONFIG.base_url_override,
     )
 
-    # Buyer thresholds: buyers are more price-sensitive (want low prices).
-    # Lower price_idx = lower offered price = better for buyer.
-    if agent_role == "buyer":
-        action_type = (
-            ActionType.ACCEPT_OFFER
-            if price_idx <= 5
-            else ActionType.COUNTER_OFFER
-            if price_idx <= 7
-            else ActionType.REJECT_OFFER
-        )
-    else:
-        # When acting as seller (fallback), use seller-like thresholds.
+    # Seller thresholds: sellers are less price-sensitive (want to sell compute).
+    if agent_role == "seller":
         action_type = (
             ActionType.ACCEPT_OFFER
             if price_idx <= 7
             else ActionType.COUNTER_OFFER
             if price_idx == 8
+            else ActionType.REJECT_OFFER
+        )
+    else:
+        action_type = (
+            ActionType.ACCEPT_OFFER
+            if price_idx <= 6
+            else ActionType.COUNTER_OFFER
+            if price_idx == 7
             else ActionType.REJECT_OFFER
         )
